Merges with master, cleans up namespaces, removes CUDAContext class
