+++ aten/src/ATen/cuda/CUDAContext.h
@@ -1,0 +1,107 @@
+#pragma once
+
+#include "ATen/Context.h"
+#include "ATen/cuda/CUDAStream.h"
+#include "ATen/cuda/Exceptions.h"
+#include "ATen/ATenGeneral.h"
+
+#include "THC/THCGeneral.h"
+
+#include <cstdint>
+
+#include "cuda_runtime_api.h"
+#include "cusparse.h"
+
+namespace at { namespace cuda { 
+
+/*
+A common CUDA interface for ATen.
+
+This interface is distinct from CUDAHooks, which defines an interface that links 
+to both CPU-only and CUDA builds. That interface is intended for runtime dispatch
+and should be used from files that are included in both CPU-only and CUDA builds.
+
+CUDAContext, on the other hand, should be preferred by files only included in
+CUDA builds. It is intended to expose CUDA functionality in a consistent
+manner.
+
+This means there is some overlap between the CUDAContext and CUDAHooks, but
+the choice of which to use is simple: use CUDAContext when in a CUDA-only file,
+use CUDAHooks otherwise.
+
+Note that CUDAContext simply defines an interface with no associated class.
+It is expected that the modules whose functions compose this interface will
+manage their own state. There is only a single CUDA context/state.
+*/
+
+/* Device info */
+AT_API inline int64_t getNumGPUs() {
+  int count;
+  AT_CUDA_CHECK(cudaGetDeviceCount(&count));
+  return count;
+}
+
+AT_API inline int64_t current_device() {
+  int cur_device;
+  AT_CUDA_CHECK(cudaGetDevice(&cur_device));
+  return cur_device;
+}
+
+AT_API inline cudaDeviceProp* getCurrentDeviceProperties() {
+  return THCState_getCurrentDeviceProperties(at::globalContext().getTHCState());
+}
+
+AT_API inline cudaDeviceProp* getDeviceProperties(int64_t device) {
+  return THCState_getDeviceProperties(at::globalContext().getTHCState(), (int)device);
+}
+
+/* Streams */
+AT_API inline CUDAStream createCUDAStream() {
+  return detail::CUDAStream_createAndRetainWithOptions(
+    CUDAStream::DEFAULT_FLAGS
+  , CUDAStream::DEFAULT_PRIORITY
+  );
+}
+
+AT_API inline CUDAStream createCUDAStreamWithOptions(int32_t flags, int32_t priority) {
+  return detail::CUDAStream_createAndRetainWithOptions(flags, priority);
+}
+
+AT_API inline CUDAStream getDefaultCUDAStream() {
+  return detail::CUDAStream_getDefaultStream();
+}
+
+AT_API inline CUDAStream getDefaultCUDAStreamOnDevice(int64_t device) {
+  return detail::CUDAStream_getDefaultStreamOnDevice(device);
+}
+
+AT_API inline CUDAStream getCurrentCUDAStream() {
+  return detail::CUDAStream_getAndRetainCurrentStream();
+}
+
+AT_API inline CUDAStream getCurrentCUDAStreamOnDevice(int64_t device) {
+  return detail::CUDAStream_getAndRetainCurrentStreamOnDevice(device);
+}
+
+AT_API inline void setCurrentCUDAStream(CUDAStream stream) {
+  return detail::CUDAStream_setStream(stream.internals());
+}
+
+AT_API inline void setCurrentCUDAStreamOnDevice(int64_t device, CUDAStream stream) {
+  return detail::CUDAStream_setStreamOnDevice(device, stream.internals());
+}
+
+AT_API inline void uncheckedSetCurrentCUDAStreamOnDevice(int64_t device, CUDAStream stream) {
+  return detail::CUDAStream_uncheckedSetStreamOnDevice(device, stream.internals());
+}
+
+/* Handles */
+#ifndef __HIP_PLATFORM_HCC__
+  AT_API inline cusparseHandle_t getCurrentCUDASparseHandle() {
+    return THCState_getCurrentSparseHandle(at::globalContext().getTHCState());
+  }
+#endif
+
+
+} // namespace cuda
+} // namespace at
