+++ aten/src/ATen/cuda/detail/CUDAHooks.cpp
@@ -49,37 +49,12 @@ void unchecked_set_device(int32_t device) {
   (void)return_code;
 }
 
-void cuda_stream_create_with_priority(
-  cudaStream_t* pStream
-, int32_t flags
-, int32_t priority) {
-#ifndef __HIP_PLATFORM_HCC__
-  check_status(cudaStreamCreateWithPriority(pStream, flags, priority));
-#else
-  check_status(cudaStreamCreateWithFlags(pStream, flags));
-#endif
-}
-
-void cuda_stream_destroy(cudaStream_t stream) {
-  check_status(cudaStreamDestroy(stream));
-}
-
-void unchecked_cuda_stream_destroy(cudaStream_t stream) {
-  const auto return_code = cudaStreamDestroy(stream);
-  (void)return_code;
-}
-
 struct DynamicCUDAInterfaceSetter {
   DynamicCUDAInterfaceSetter() {
     using at::detail::DynamicCUDAInterface;
     DynamicCUDAInterface::set_device = set_device;
     DynamicCUDAInterface::get_device = get_device;
     DynamicCUDAInterface::unchecked_set_device = unchecked_set_device;
-    DynamicCUDAInterface::cuda_stream_create_with_priority =
-        cuda_stream_create_with_priority;
-    DynamicCUDAInterface::cuda_stream_destroy = cuda_stream_destroy;
-    DynamicCUDAInterface::unchecked_cuda_stream_destroy =
-        unchecked_cuda_stream_destroy;
   }
 };
 
@@ -123,21 +98,6 @@ bool CUDAHooks::hasCuDNN() const {
   return AT_CUDNN_ENABLED();
 }
 
-#ifndef __HIP_PLATFORM_HCC__
-cusparseHandle_t CUDAHooks::getCurrentCUDASparseHandle(THCState* thc_state) const {
-  return THCState_getCurrentSparseHandle(thc_state);
-}
-#endif
-struct cudaDeviceProp* CUDAHooks::getCurrentDeviceProperties(
-    THCState* thc_state) const {
-  return THCState_getCurrentDeviceProperties(thc_state);
-}
-struct cudaDeviceProp* CUDAHooks::getDeviceProperties(
-    THCState* thc_state,
-    int device) const {
-  return THCState_getDeviceProperties(thc_state, device);
-}
-
 int64_t CUDAHooks::current_device() const {
   int device;
   cudaError_t err = cudaGetDevice(&device);
@@ -162,7 +122,7 @@ bool CUDAHooks::compiledWithCuDNN() const {
 bool CUDAHooks::supportsDilatedConvolutionWithCuDNN() const {
 #if AT_CUDNN_ENABLED()
   cudaDeviceProp* prop =
-      getCurrentDeviceProperties(globalContext().getTHCState());
+      THCState_getCurrentDeviceProperties(globalContext().getTHCState());
   // NOTE: extra parenthesis around numbers disable clang warnings about
   // dead code
   return (
