+++ aten/src/ATen/cuda/detail/CUDAHooks.h
@@ -14,11 +14,6 @@ struct CUDAHooks : public at::CUDAHooksInterface {
   std::unique_ptr<Generator> initCUDAGenerator(Context*) const override;
   bool hasCUDA() const override;
   bool hasCuDNN() const override;
-#ifndef __HIP_PLATFORM_HCC__
-  cusparseHandle_t getCurrentCUDASparseHandle(THCState*) const override;
-#endif
-  struct cudaDeviceProp* getCurrentDeviceProperties(THCState*) const override;
-  struct cudaDeviceProp* getDeviceProperties(THCState*, int device) const override;
   int64_t current_device() const override;
   Allocator* getPinnedMemoryAllocator() const override;
   void registerCUDATypes(Context*) const override;
