+++ aten/src/ATen/native/cuda/SpectralOps.cu
@@ -1,4 +1,5 @@
 #include "ATen/ATen.h"
+#include "ATen/cuda/CUDAContext.h"
 #include "ATen/Config.h"
 #include "ATen/Dispatch.h"
 #include "ATen/Utils.h"
@@ -104,7 +105,7 @@ static void _fft_fill_with_conjugate_symmetry_(Tensor& input,
   // copy
   int64_t n = input.numel() / size_last_dim * (size_last_dim - last_dim_start_slice);
 
-  cudaStream_t stream = globalContext().getCurrentCUDAStream();
+  cudaStream_t stream = at::cuda::getCurrentCUDAStream();
   auto allocator = THCThrustAllocator(globalContext().lazyInitCUDA());
   auto policy = thrust::cuda::par(allocator).on(stream);
   AT_DISPATCH_FLOATING_TYPES_AND_HALF(input.type(), "_fft_fill_with_conjugate_symmetry_", [&] {
@@ -180,7 +181,7 @@ static inline Tensor _run_cufft(
   auto output = input.type().tensor(output_sizes);
 
   // set to current stream
-  CUFFT_CHECK(cufftSetStream(plan, ctx.getCurrentCUDAStream()));
+  CUFFT_CHECK(cufftSetStream(plan, at::cuda::getCurrentCUDAStream()));
 
   auto ws = ctx.getType(at::Backend::CUDA, at::ScalarType::Byte).tensor({ config.workspace_size() });
   CUFFT_CHECK(cufftSetWorkArea(plan, ws.data_ptr()));
