+++ aten/src/ATen/native/cudnn/RNN.cpp
@@ -5,6 +5,7 @@
 #include <ATen/MatrixRef.h>
 #include <ATen/NativeFunctions.h>
 #include <ATen/cuda/CUDAConfig.h>
+#include <ATen/cuda/Exceptions.h>
 
 #if !AT_CUDNN_ENABLED()
