+++ aten/src/ATen/native/sparse/cuda/SparseCUDABlas.cu
@@ -1,6 +1,7 @@
 #include <ATen/native/sparse/cuda/SparseCUDABlas.cuh>
 #include <ATen/Error.h>
 #include <ATen/Context.h>
+#include <ATen/cuda/CUDAContext.h>
 
 #include <TH/THGeneral.h>
 
@@ -60,8 +61,8 @@ inline void CUSPARSE_CHECK(cusparseStatus_t status)
 }
 
 inline cusparseHandle_t setCUDASparseStream() {
-  cusparseHandle_t handle = globalContext().getCurrentCUDASparseHandle();
-  cusparseSetStream(handle, globalContext().getCurrentCUDAStream());
+  cusparseHandle_t handle = at::cuda::getCurrentCUDASparseHandle();
+  cusparseSetStream(handle, at::cuda::getCurrentCUDAStream());
   return handle;
 }
